# perf(ts): reuse Reader whilst parsing records