--- a/typescript/core/src/McapIndexedReader.ts
+++ b/typescript/core/src/McapIndexedReader.ts
@@ -2,10 +2,10 @@ import { crc32, crc32Final, crc32Init, crc32Update } from "@foxglove/crc";
 import Heap from "heap-js";
 
 import { ChunkCursor } from "./ChunkCursor";
+import Reader from "./Reader";
 import { MCAP_MAGIC } from "./constants";
 import { parseMagic, parseRecord } from "./parse";
 import { DecompressHandlers, IReadable, TypedMcapRecords } from "./types";
-import Reader from "./Reader";
 
 type McapIndexedReaderArgs = {
   readable: IReadable;