… typescript/core/src/McapWriter.test.ts
… typescript/core/src/StreamBuffer.test.ts
… typescript/core/src/StreamBuffer.ts